+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
-* June 30, 2014
+* July 17, 2014
+
+  - Added width parameter in ``add_label()``.
 
   - Added private members in all documentation.